+++ include/hep/mc.hpp
@@ -33,6 +33,7 @@
 #include "hep/mc/multi_channel_max_difference.hpp"
 #include "hep/mc/multi_channel_point.hpp"
 #include "hep/mc/multi_channel_result.hpp"
+#include "hep/mc/multi_channel_weight_info.hpp"
 #include "hep/mc/plain.hpp"
 #include "hep/mc/plain_result.hpp"
 #include "hep/mc/projector.hpp"